Renamed change-log-data-service to change-log-store
#xxx
