+++ waltz-ng/client/data-types/services/data-type-view-data.js
@@ -146,7 +146,7 @@ service.$inject = [
     '$q',
     'ApplicationStore',
     'AppCapabilityStore',
-    'ChangeLogDataService',
+    'ChangeLogStore',
     'DataTypeUsageStore',
     'RatingStore',
     'DataTypeService',
